Conflicts:
	compass_ae_console/lib/compass_ae_console/version.rb
	erp_app/lib/erp_app/version.rb
	erp_base_erp_svcs/app/models/party.rb
	erp_base_erp_svcs/db/migrate/20080805000020_base_erp_services.rb
	erp_base_erp_svcs/lib/erp_base_erp_svcs/version.rb
	erp_commerce/lib/erp_commerce/version.rb
	erp_invoicing/lib/erp_invoicing/version.rb
	erp_orders/lib/erp_orders/version.rb
	erp_products/lib/erp_products/version.rb
	erp_tech_svcs/config/initializers/erp_tech_svcs.rb
	erp_tech_svcs/erp_tech_svcs.gemspec
	erp_tech_svcs/lib/erp_tech_svcs/file_support/s3_manager.rb
	erp_tech_svcs/lib/erp_tech_svcs/version.rb
	erp_txns_and_accts/lib/erp_txns_and_accts/version.rb
	knitkit/knitkit.gemspec
	knitkit/lib/knitkit/version.rb
	knitkit/public/javascripts/erp_app/desktop/applications/knitkit/image_assets_panel.js

+++ erp_commerce/lib/erp_commerce/active_merchant_wrappers/brain_tree_gateway_wrapper.rb
@@ -27,7 +27,7 @@ module ErpCommerce
           })
 
         if credit_card_result.valid?
-          cents = (amount.to_f * 100)
+          cents = (amount.to_d * 100)
           response = gateway.purchase(cents, credit_card_result)
 
           if response.success?
